requirements.txt
django==1.10
django-webpack-loader==0.3.3
djangorestframework==3.4.7
wheel==0.29.0
gunicorn==19.6.0
requests==2.11.1
